requirements.txt
streamlit
azure-ai-vision
python-dotenv
pillow
